Fix compiler warning from newer GCC (Issue #5533)

+++ ppdc/testcatalog.cxx
@@ -1,9 +1,10 @@
 //
 // Test program for message catalog class.
 //
-// Copyright 2008 by Apple Inc.
+// Copyright © 2008-2019 by Apple Inc.
 //
-// Licensed under Apache License v2.0.  See the file "LICENSE" for more information.
+// Licensed under Apache License v2.0.  See the file "LICENSE" for more
+// information.
 //
 
 //
@@ -34,7 +35,7 @@ main(int  argc,				// I - Number of command-line arguments
   // Scan the command-line...
   catalog = new ppdcCatalog(NULL, argv[1]);
 
-  printf("%s: %d messages\n", argv[1], catalog->messages->count);
+  printf("%s: %u messages\n", argv[1], (unsigned)catalog->messages->count);
 
   for (m = (ppdcMessage *)catalog->messages->first();
        m;
